Don't error out on read_body fail
If the server sends a malformed body with the response, parse as much
as possible and then display the results as is.

This is the first step of a larger refactor to allow http_client to be
more permissive so that the user receives as much information back as
possible without triggering a silent fatal error.