Initial implementation of HTTP caching with Cache-Control
This is intentionally as simple as possible. We could fully implement
all the different cache-control types on a per resource basis, plus
even the last-modified/expires style caching, but that would simply be
added code with dubious performance gains for most workloads. If
you're really in the situation where HTTP caching could provide a
significant performance boost, why in the love of god are you running
a shell script as a web server??