Remove IP Match module
Cool concept, but didn't actually work in practice beyond being able
to filter IPs on a local network. The problem is that ztcp doesn't
necessarily record the IP address of the client in the session
table. Instead, it saves the "fully qualified domain name of the peer,
if available, else an IP address." It helps to properly read the
documentation, kids.

Plus, there was a horrible bug in parse_ztcp that was hidden by virtue
of our unit tests running on localhost. In a nutshell, client_ip was
being set to the local host name, usually "0.0.0.0", due to a wrong
index position --- fix coming later since it still affects logging.